+++ erigon-lib/commitment/hex_patricia_hashed.go
@@ -44,7 +44,7 @@ import (
 
 	"github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/common/length"
-	"github.com/erigontech/erigon-lib/rlp"
+	rlp "github.com/erigontech/erigon-lib/rlp2"
 )
 
 // keccakState wraps sha3.state. In addition to the usual hash methods, it also supports
