+++ erigon-lib/rlp2/parse.go
@@ -26,10 +26,9 @@ import (
 )
 
 var (
-	ErrBase          = errors.New("rlp")
-	ErrParse         = fmt.Errorf("%w parse", ErrBase)
-	ErrDecode        = fmt.Errorf("%w decode", ErrBase)
-	ErrUnexpectedEOF = fmt.Errorf("%w EOF", ErrBase)
+	ErrBase   = errors.New("rlp")
+	ErrParse  = fmt.Errorf("%w parse", ErrBase)
+	ErrDecode = fmt.Errorf("%w decode", ErrBase)
 )
 
 func IsRLPError(err error) bool { return errors.Is(err, ErrBase) }
@@ -37,8 +36,8 @@ func IsRLPError(err error) bool { return errors.Is(err, ErrBase) }
 // BeInt parses Big Endian representation of an integer from given payload at given position
 func BeInt(payload []byte, pos, length int) (int, error) {
 	var r int
-	if pos+length > len(payload) {
-		return 0, ErrUnexpectedEOF
+	if pos+length >= len(payload) {
+		return 0, fmt.Errorf("%w: unexpected end of payload", ErrParse)
 	}
 	if length > 0 && payload[pos] == 0 {
 		return 0, fmt.Errorf("%w: integer encoding for RLP must not have leading zeros: %x", ErrParse, payload[pos:pos+length])
