--- erigon-lib/rlp2/readme.md
@@ -1,11 +1,0 @@
-## rlp
-
-
-TERMINOLOGY:
-
-```
-RLP string = "Blob" // this is so we don't conflict with existing go name for String
-RLP list = "List" // luckily we can keep using list name since go doesn't use it
-RLP single byte number = "Decimal" // for numbers from 1-127. a special case
-```
-
